+++ src/main/java/com/bob_senior/bob_server/domain/Chat/SessionAndClientRecord.java
@@ -9,5 +9,5 @@ public class SessionAndClientRecord {
 
     private String sessionId;
 
-    private Integer userIdx;
+    private Long userIdx;
 }
