+++ src/main/java/com/bob_senior/bob_server/domain/Chat/entity/ChatParticipant.java
@@ -16,7 +16,7 @@ public class ChatParticipant {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer chatParticipantIdx;
+    private Long chatParticipantIdx;
 
     @Embedded
     private ChatNUser chatNUser;
