+++ src/main/java/com/bob_senior/bob_server/domain/Chat/entity/ChatRoom.java
@@ -8,7 +8,7 @@ import javax.persistence.Id;
 public class ChatRoom {
 
     @Id
-    private Integer chatRoomIdx;
+    private Long chatRoomIdx;
 
     @Column(name = "chatRoomName")
     private String chatRoomName;
