+++ src/main/java/com/bob_senior/bob_server/domain/Post/entity/PhotoId.java
@@ -17,7 +17,7 @@ import java.io.Serializable;
 public class PhotoId implements Serializable {
 
     @Column(name = "postIdx")
-    private Integer postIdx;
+    private Long postIdx;
 
     @Column(name = "postPhotoUrl")
     private String postPhotoUrl;
