+++ src/main/java/com/bob_senior/bob_server/domain/Post/entity/TagId.java
@@ -16,6 +16,6 @@ public class TagId implements Serializable {
 
 
     @Column(name = "postIdx")
-    private Integer postIdx;
+    private Long postIdx;
 
 }
