+++ src/main/java/com/bob_senior/bob_server/domain/user/DetermineFriendshipDTO.java
@@ -4,9 +4,9 @@ import lombok.Getter;
 
 @Getter
 public class DetermineFriendshipDTO {
-    private Integer userIdx;
+    private Long userIdx;
 
-    private Integer targetIdx;
+    private Long targetIdx;
 
     private boolean accept;
 }
