+++ src/main/java/com/bob_senior/bob_server/domain/user/entity/Friendship.java
@@ -15,7 +15,7 @@ public class Friendship {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer friendshipIdx;
+    private Long friendshipIdx;
 
 
     @Embedded
