+++ src/main/java/com/bob_senior/bob_server/domain/vote/TerminateVoteDTO.java
@@ -7,7 +7,7 @@ import lombok.Setter;
 @Setter
 public class TerminateVoteDTO {
 
-    private Integer terminatorIdx;
-    private Integer voteIdx;
+    private Long terminatorIdx;
+    private Long voteIdx;
 
 }
