+++ src/main/java/com/bob_senior/bob_server/domain/vote/entity/Vote.java
@@ -15,13 +15,13 @@ public class Vote {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer voteIdx;
+    private Long voteIdx;
 
     @Column
-    private Integer postIdx;
+    private Long postIdx;
 
     @Column
-    private Integer creatorIdx;
+    private Long creatorIdx;
 
     @Column
     private String title;
