+++ src/main/java/com/bob_senior/bob_server/domain/vote/entity/VoteParticipated.java
@@ -13,11 +13,11 @@ import javax.persistence.*;
 public class VoteParticipated {
 
     @Id
-    private Integer voteParticipatedIdx;
+    private Long voteParticipatedIdx;
 
 
     @Column
-    private Integer userIdx;
+    private Long userIdx;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "voteIdx")
