+++ src/main/java/com/bob_senior/bob_server/repository/PostPhotoRepository.java
@@ -3,10 +3,10 @@ package com.bob_senior.bob_server.repository;
 import com.bob_senior.bob_server.domain.Post.entity.PostPhoto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PostPhotoRepository extends JpaRepository<PostPhoto, Integer> {
+public interface PostPhotoRepository extends JpaRepository<PostPhoto, Long> {
 
 
-    PostPhoto findPostPhotoByPost_PostIdx(Integer postIdx);
+    PostPhoto findPostPhotoByPost_PostIdx(Long postIdx);
