+++ src/main/java/com/bob_senior/bob_server/repository/VoteRecordRepository.java
@@ -6,12 +6,12 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface VoteRecordRepository extends JpaRepository<VoteRecord, Integer> {
+public interface VoteRecordRepository extends JpaRepository<VoteRecord, Long> {
 
     VoteRecord findVoteRecordByVoteId(VoteId id);
 
-    List<VoteRecord> findAllByVoteId_VoteIdx(Integer voteIdx);
+    List<VoteRecord> findAllByVoteId_VoteIdx(Long voteIdx);
 
-    VoteRecord findFirstByVoteId_VoteIdxOrderByCountDesc(Integer voteIdx);
+    VoteRecord findFirstByVoteId_VoteIdxOrderByCountDesc(Long voteIdx);
 
 }
